fix(react-ts): move react deps to webDependencies
This was necessary for me to see the app load on the page. A blank screen would be shown otherwise, with an error in the console:

```
TypeError: Z is undefinedreact-dom.js:340:200
    React 3
        reactDom_production_min
        reactDom_production_min
        reactDom_production_min
    createCommonjsModule http://localhost:3000/web_modules/common/_commonjsHelpers-f5462f22.js:4
    <anonymous> React
    InnerModuleEvaluation self-hosted:1604
    InnerModuleEvaluation self-hosted:1593
    evaluation self-hosted:1558
```

The removal of the extension in `index.tsx`  was necessary for `lintall:tsc` not to complain.

I believe a similar fix should be done in the React template.